Fix collections SSR
Extract common functionality to a composable
